Project screen and color changes
I had doubts about the colours, so I had an idea about using colors
only where we need them. In Chimps section, to welcome the users and in
Projects section using the color of the project (or a similar one)
we’re talking about.

The changes are:
Clients now is brown and white, just like Team section.
Projects has now 4 projects, each one with its own color.

Hope you like it. :)

To the end, it’s missing only the projects texts and the contact
section.