Adjustments to Desktop screens (read the description of this commit please)
This scss file will be the name changed.
Right now for reasons of easy test I just set this styles to be applied
in >1100px resolutions. Next step is filter the style who apply only in
landscape tablets and separate it from screen1024.scss to a new file.
After that the styles on this file will be migrated to screen1024.scss.